requirements.txt
anomalib
torch
torchvision
openvino
opencv-python
albumentations
pandas
omegaconf
requests
fiftyone
umap-learn>=0.5
streamlit
fpdf2
ultralytics
psutil
pillow
lightning